Continue working on design
