+++ components/mdx.tsx
@@ -3,9 +3,7 @@ import Image from 'next/image';
 import type { HTMLProps, ReactNode } from 'react';
 import { Tweet } from 'react-tweet';
 import { Link } from './link';
-import { LocationCard } from './location-card';
 import { MailingList } from './mailing-list';
-import { Projects } from './projects';
 import { Social } from './social';
 import { Stack } from './stack';
 import { Video } from './video';
@@ -112,8 +110,6 @@ export const Mdx = ({ code }: MdxProperties) => (
       Social,
       Stack,
       MailingList,
-      LocationCard,
-      Projects,
     }}
   />
 );
