+++ content/pages/home.mdx
@@ -3,9 +3,9 @@ title: Hayden Bleasel
 description: I'm an Australian Design Engineer living in 🇺🇸 San Francisco, California.
 ---
 
-I'm an Australian Design Engineer living in <LocationCard timezone="America/Los_Angeles">🇺🇸 San Francisco, California</LocationCard>. I love creating beautiful software that delights users and reimagines the way we interact with technology. I'm also an avid drone pilot, gamer and open source maintainer.
+I'm an Australian Design Engineer living in 🇺🇸 San Francisco, California. I love creating beautiful software that delights users and reimagines the way we interact with technology. I'm also an avid drone pilot, gamer and open source maintainer.
 
-I'm currently on the Developer Experience team at [Vercel](https://vercel.com) where I focus on building our community, sharing knowledge and helping developers build better products. After hours, I'm exploring new ideas (particularly AI-native apps) and building open source software. If you like my OSS work, you can [sponsor me](https://github.com/sponsors/haydenbleasel).
+I'm currently on the Developer Experience team at Vercel where I focus on building our community, sharing knowledge and helping developers build better products. After hours, I'm exploring new ideas (particularly AI-native apps) and building open source software. If you like my OSS work, you can [sponsor me](https://github.com/sponsors/haydenbleasel).
 
 I maintain a small mailing list where I share infrequent updates on my projects and what's new. Join 2000+ readers if you're interested!
