+++ content/pages/projects.mdx
@@ -9,10 +9,13 @@ One of my earlier successes was a tool for learning, improving and generating co
 
 I later built [next-forge](https://www.next-forge.com/) — a production-grade Turborepo template for Next.js apps. I released it for free in a market full of paid boilerplates and grew it to nearly 6,000 stars on GitHub, fostering a community of contributors and partners around it. It got some great coverge after Guillermo Rauch [posted it](https://x.com/rauchg/status/1853171412766466119), including a [YouTube review](https://youtu.be/aHZz_6NhQkQ) by Josh tried coding. Vercel [acquired](https://x.com/haydenbleasel/status/1929625673586598148) next-forge in 2025 as part of my hiring.
 
-My most recent successful project is Orate &mdash; the AI toolkit for speech. It's a library for synthesizing speech, transcribing audio, converting voices and isolating audio. After growing it to around 400 stars on GitHub, Orate was merged into the Vercel AI SDK in 2025 as [`transcribe()`](https://x.com/aisdk/status/1909603418639155460) and [`generateSpeech()`](https://x.com/aisdk/status/1911763995272626504) functions.
+My most recent successful project is Orate &mdash; the AI toolkit for speech. It's a library for synthesizing speech, transcribing audio, converting voices and isolating audio. After growing it to around 400 stars on GitHub, Orate was merged into the Vercel [AI SDK](https://ai-sdk.dev/) in 2025 as [`transcribe()`](https://x.com/aisdk/status/1909603418639155460) and [`generateSpeech()`](https://x.com/aisdk/status/1911763995272626504) functions.
 
-Many of my projects are still active or in progress, including:
+Some of my projects are still active or in progress, including:
 
-<Projects />
+- [Tersa](https://www.tersa.ai/) - An open source canvas for building AI workflows.
+- [Ultracite](https://www.ultracite.dev/) - Robust linting config for TypeScript apps, built on Biome.
+- [Kibo UI](https://www.kibo-ui.com/) - Composable, open-source shadcn/ui components for React.
+- [Eververse](https://www.eververse.ai/) - The open source product management platform.
 
-[^1]: I've also worked with Advancell, Airwallex, Audience Republic, Baraja, Brighte, Comfort Delgro, Elevio, Faethm, Flaunter, Flirtey, Futrli, Grow, Inventia, Kerbly, Lightswap, Lookahead, Notiv, Perlin, Pursuited, Resolution Collective, Rezi, Ribit, Shippit, Siesta Campers, Simply Wall St, Snug, Space Machines Company, Spaceship, Tank Stream Ventures, Teleqo Technologies, Tyro Payments, UpGuard, UTS, Zibbet and Zookal.
+[^1]: Refraction was used by Accenture, AKQA, Amazon, Bentley, Canva, CapGemini, Cisco, Experian, ExpressVPN, Google, H&M, Ikea, John Deere, Nespresso, PandaDoc, Qantas, Rakuten, Red Bull, Repl.it, Roblox, Softbank, Sega, Tiktok, Uber, Washington Post, Wix, Yahoo, Zoho and ZoomInfo.
