+++ content/pages/work.mdx
@@ -7,8 +7,10 @@ My previous role was Chief Product Officer at [Corellium](https://www.corellium.
 
 Before that, I was Director of [Jellypepper](https://jellypepper.com/), an award-winning digital agency for disruptive startups. We worked with early stage tech companies in niche areas such as self-driving car tech, AI, biotech, crypto, drone delivery, cybersecurity and even orbital (outer space) logistics. Jellypepper was [acquired](https://raw.studio/blog/raw-studio-acquires-jellypepper-to-expand-its-reach-to-the-startup-ecosystem/) by Raw Studio in 2024.
 
-Earlier in my career, I was the Head of Product and Design at [Spaceship](https://www.spaceship.com.au/), a leading Australian investing platform later [acquired by eToro](https://www.spaceship.com.au/news/etoro-completes-acquisition-spaceship/); and did a product design internship at [Palantir](https://www.palantir.com/). Through various contracts, I was also fortunate enough to work with Australian Ethical, Canva, Clipchamp, ESLint, Google, National Geographic, Nike, Node.js, R/GA, Timberland, Toyota, Westfield and many more brilliant companies[^1].
+Earlier in my career, I was the Head of Product and Design at [Spaceship](https://www.spaceship.com.au/) (a leading Australian investing platform later [acquired](https://www.spaceship.com.au/news/etoro-completes-acquisition-spaceship/) by eToro); contracted for [R/GA](https://www.rga.com/) and did a product design internship at [Palantir](https://www.palantir.com/) in Palo Alto, California.
 
-I've also enjoyed working with open source organizations. You may have seen some of my work out there — [ESLint](https://eslint.org/blog/2022/08/redesigning-eslint/), [Node.js](https://nodejs.org/en/blog/announcements/diving-into-the-nodejs-website-redesign) and [SWC plugins](https://plugins.swc.rs/) for example.
+I've also had the pleasure of working with open source organizations. You may have seen some of my work out there — [ESLint](https://eslint.org/blog/2022/08/redesigning-eslint/), [Node.js](https://nodejs.org/en/blog/announcements/diving-into-the-nodejs-website-redesign) and [SWC plugins](https://plugins.swc.rs/) for example.
 
-[^1]: I've also worked with Advancell, Airwallex, Audience Republic, Baraja, Brighte, Comfort Delgro, Elevio, Faethm, Flaunter, Flirtey, Futrli, Grow, Inventia, Kerbly, Lightswap, Lookahead, Notiv, Perlin, Pursuited, Resolution Collective, Rezi, Ribit, Shippit, Siesta Campers, Simply Wall St, Snug, Space Machines Company, Spaceship, Tank Stream Ventures, Teleqo Technologies, Tyro Payments, UpGuard, UTS, Zibbet and Zookal.
+I've also been fortunate enough to work on a contracting basis with Australian Ethical, Canva, Clipchamp, ESLint, Google, National Geographic, Nike, Node.js, R/GA, Timberland, Toyota, Westfield and many more brilliant companies[^1].
+
+[^1]: Advancell, Airwallex, Audience Republic, Baraja, Brighte, Comfort Delgro, Elevio, Faethm, Flaunter, Flirtey, Futrli, Grow, Inventia, Kerbly, Lightswap, Lookahead, Notiv, Perlin, Profound, Pursuited, Resolution Collective, Rezi, Ribit, Shippit, Siesta Campers, Simply Wall St, Snug, Space Machines Company, Spaceship, Tank Stream Ventures, Teleqo Technologies, Tyro Payments, UpGuard, UTS, Zibbet and Zookal.
