--- lib/projects.ts
@@ -1,22 +1,0 @@
-export const projects = [
-  {
-    name: 'Tersa',
-    description: 'An open source canvas for building AI workflows.',
-    link: 'https://www.tersa.ai/',
-  },
-  {
-    name: 'Ultracite',
-    description: 'Robust linting config for TypeScript apps, built on Biome.',
-    link: 'https://www.ultracite.dev/',
-  },
-  {
-    name: 'Kibo UI',
-    description: 'Composable, open-source shadcn/ui components for React.',
-    link: 'https://www.kibo-ui.com/',
-  },
-  {
-    name: 'Eververse',
-    description: 'The open source product management platform.',
-    link: 'https://www.eververse.ai/',
-  },
-];
